server/requirements.txt
boto3 == 1.26.153
Django == 4.2.1
fastapi == 0.97.0
pydantic == 1.10.9
pymongo == 4.3.3
python-dotenv == 1.0.0
'uvicorn[standard]'
